✨ feat: 로그인 외부 API 격리 설정 추가
