+++ src/main/java/com/mallang/mallang_backend/domain/member/service/main/MemberService.java
@@ -11,10 +11,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface MemberService {
 
-    boolean existsByPlatformId(String platformId);
-
-    Long signupByOauth(String platformId, String email, String nickname, String profileImage, LoginPlatform loginPlatform);
-
     Member getMemberByPlatformId (String platformId);
 
     String getRoleName(Long memberId);
